chore: apply lint fixes
