+++ packages/fit/src/adapters/krd-to-fit/krd-to-fit-step-count.helpers.ts
@@ -1,0 +1,22 @@
+import type { RepetitionBlock, WorkoutStep } from "@kaiord/core";
+import { isRepetitionBlock } from "@kaiord/core";
+
+/**
+ * Counts the number of FIT-valid workout steps from a list of KRD steps.
+ *
+ * Each repetition block expands to its inner steps plus one synthetic
+ * repeat step in the FIT message stream.
+ */
+export const countValidSteps = (
+  steps: Array<WorkoutStep | RepetitionBlock>
+): number => {
+  let count = 0;
+  for (const step of steps) {
+    if (isRepetitionBlock(step)) {
+      count += step.steps.length + 1;
+    } else {
+      count += 1;
+    }
+  }
+  return count;
+};
